@@ -2,102 +2,57 @@
   "name": "communication-server",
   "version": "2.0.1",
   "scripts": {
-    "ng": "ng",
-    "format": "prettier --write \"**/*.ts\"",
-    "format:check": "prettier-check \"**/*.ts\"",
-    "ng:start": "ng serve",
-    "e2e": "ng e2e --configuration=ci",
-    "build": "ng build",
-    "lint": "./node_modules/.bin/tslint -c tslint.json 'src/**/*.ts'",
+    "lint": "tslint -c tslint.json './src/**/*.ts' './tests/**/*.ts'",
+    "format": "prettier --write \"./src/**/*.ts\" \"./tests/**/*.ts\"",
+    "format:check": "prettier-check \"./src/**/*.ts\" \"./tests/**/*.ts\"",
     "start:server": "nodemon",
     "start:debug": "nodemon --config nodemon-debug.json",
-    "start:client": "ng serve --proxy-config proxy.config.json",
-    "start:client-dev": "ng build --watch=true",
-    "start:dev": "concurrently --kill-others-on-fail \"npm run start:server\" \"npm run start:client\"",
-    "prestart:prod": "rm -rf dist && ng build --prod && tsc -p src/server/tsconfig.server.json",
-    "start:prod": "node dist/main.js",
-    "start:hmr": "node dist/server",
+    "build": "rm -rf dist && tsc -p src/tsconfig.build.json",
+    "start:prod": "node dist/out-tsc/main.js",
     "test:server": "jest",
     "test:cov": "jest --coverage",
-    "test:e2e": "export NODE_ENV=test-e2e && jest --config ./tests/jest-e2e.json --detectOpenHandles",
-    "test:client": "ng test --watch=false --no-progress",
-    "webpack": "webpack --config webpack.config.js"
+    "test:e2e": "export NODE_ENV=test-e2e && jest --config ./tests/jest-e2e.json --detectOpenHandles"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.10",
-    "@angular/cdk": "^7.3.5",
-    "@angular/common": "^7.2.10",
-    "@angular/compiler": "^7.2.10",
-    "@angular/core": "^7.2.10",
-    "@angular/flex-layout": "^7.0.0-beta.22",
-    "@angular/forms": "^7.2.10",
-    "@angular/http": "^7.2.10",
-    "@angular/material": "^7.3.5",
-    "@angular/platform-browser": "^7.2.10",
-    "@angular/platform-browser-dynamic": "^7.2.10",
-    "@angular/router": "^7.2.10",
-    "@nestjs/common": "^6.0.2",
-    "@nestjs/core": "^6.0.2",
+    "@nestjs/common": "^6.0.4",
+    "@nestjs/core": "^6.0.4",
     "@nestjs/cqrs": "^6.0.0",
-    "@nestjs/microservices": "^6.0.2",
-    "@nestjs/platform-express": "^6.0.2",
+    "@nestjs/microservices": "^6.0.4",
+    "@nestjs/platform-express": "^6.0.4",
     "@nestjs/swagger": "^3.0.2",
     "@nestjs/typeorm": "^6.0.0",
-    "angular-oauth2-oidc": "^5.0.2",
-    "aws-sdk": "^2.427.0",
+    "aws-sdk": "^2.431.0",
     "class-transformer": "^0.2.0",
     "class-validator": "^0.9.1",
-    "core-js": "^2.6.5",
-    "craft-account-manager": "^1.0.0",
     "express": "^4.16.4",
     "fastify-formbody": "^3.1.0",
-    "hammerjs": "^2.0.8",
     "joi": "^14.3.1",
     "mongodb": "^3.2.2",
-    "nodemailer": "^5.1.1",
+    "nodemailer": "^6.0.0",
     "nodemailer-express-handlebars": "^3.0.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^6.4.0",
-    "typeorm": "^0.2.15",
-    "zone.js": "^0.9.0"
+    "swagger-ui-express": "^4.0.2",
+    "typeorm": "^0.2.16"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.6",
-    "@angular/cli": "^7.3.6",
-    "@angular/compiler-cli": "^7.2.10",
-    "@angular/language-service": "^7.2.10",
-    "@nestjs/testing": "^6.0.2",
+    "@nestjs/testing": "^6.0.4",
     "@types/express": "^4.16.1",
-    "@types/jasmine": "^3.3.12",
-    "@types/jasminewd2": "~2.0.6",
     "@types/jest": "^24.0.11",
-    "@types/node": "^11.11.6",
+    "@types/node": "^11.12.1",
     "@types/supertest": "^2.0.7",
-    "codelyzer": "~4.5.0",
-    "concurrently": "^4.1.0",
-    "jasmine-core": "~3.3.0",
-    "jasmine-spec-reporter": "~4.2.1",
     "jest": "^24.5.0",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
     "nodemon": "^1.18.10",
     "prettier": "^1.16.4",
     "prettier-check": "^2.0.0",
-    "protractor": "~5.4.2",
     "supertest": "^4.0.2",
     "ts-jest": "^24.0.0",
     "ts-loader": "^5.3.3",
     "ts-node": "^8.0.3",
     "tsconfig-paths": "^3.8.0",
     "tslint": "^5.14.0",
-    "typescript": "~3.1.6",
-    "webpack": "^4.29.6",
-    "webpack-cli": "^3.3.0",
-    "webpack-node-externals": "^1.6.0"
+    "typescript": "^3.3.4000"
   },
   "jest": {
     "moduleDirectories": [
