@@ -1,7 +1,18 @@
 import { Component, OnInit } from '@angular/core';
-import { FormGroup, FormControl } from '@angular/forms';
+import { FormGroup, FormControl, FormArray, FormBuilder } from '@angular/forms';
 import { SettingsService } from './settings.service';
-import { APP_URL, ISSUER_URL } from '../constants/storage';
+import { SERVICES, ISSUER_URL } from '../constants/storage';
+import { ServiceList } from './service-list.interface';
+import { from, empty } from 'rxjs';
+import { mergeMap, toArray, map, catchError } from 'rxjs/operators';
+import { ISettings } from './settings.interface';
+import {
+  IS_DOWN,
+  CLOSE,
+  UPDATE_SUCCESSFUL,
+  ERROR_UPDATING_SERVICE_SETTINGS,
+} from '../constants/messages';
+import { MatSnackBar } from '@angular/material';
 
 @Component({
   selector: 'app-settings',
@@ -15,7 +26,11 @@ export class SettingsComponent implements OnInit {
   appURL: string;
   clientId: string;
   clientSecret: string;
-  hide: boolean = true;
+  communicationServerSystemEmailAccount: string;
+  emailAccounts: any[];
+  clientsFormGroup: FormGroup;
+  clientsFormArray: FormArray;
+  cloudStorageList: { uuid: string; name: string }[] = [];
 
   infraSettingsForm = new FormGroup({
     appURL: new FormControl(this.appURL),
@@ -28,23 +43,25 @@ export class SettingsComponent implements OnInit {
     communicationServerClientId: new FormControl(
       this.communicationServerClientId,
     ),
+    communicationServerSystemEmailAccount: new FormControl(
+      this.communicationServerSystemEmailAccount,
+    ),
   });
 
-  constructor(private settingsService: SettingsService) {}
+  constructor(
+    private settingsService: SettingsService,
+    private formBuilder: FormBuilder,
+    private snackBar: MatSnackBar,
+  ) {}
 
   ngOnInit() {
-    this.settingsService.getClientSettings().subscribe({
-      next: (clientResponse: {
-        appURL: string;
-        clientId: string;
-        clientSecret: string;
-      }) => {
-        this.appURL = clientResponse.appURL;
-        this.clientId = clientResponse.clientId;
-        this.clientSecret = clientResponse.clientSecret;
-        this.populateClientForm(clientResponse);
-      },
+    this.clientsFormGroup = this.formBuilder.group({
+      clientsFormArray: new FormArray([]),
     });
+
+    this.clientsFormArray = this.clientsFormGroup.get(
+      'clientsFormArray',
+    ) as FormArray;
 
     this.settingsService.getSettings().subscribe({
       next: (response: {
@@ -55,26 +72,97 @@ export class SettingsComponent implements OnInit {
         this.communicationServerClientId = response.communicationServerClientId;
         this.populateForm(response);
       },
+      error: error => {},
     });
+
+    this.settingsService.getSavedEmailAccount<any>().subscribe({
+      next: settings => {
+        this.communicationServerSystemEmailAccount =
+          settings.communicationServerSystemEmailAccount;
+        this.authSettingsForm.controls.communicationServerSystemEmailAccount.setValue(
+          this.communicationServerSystemEmailAccount,
+        );
+      },
+    });
+
+    this.populateServicesSettings();
+
     this.settingsService.getClientList().subscribe({
       next: (response: any[]) => {
         this.clientList = response;
       },
+      error: error => {},
     });
+
+    this.settingsService.getEmailAccounts().subscribe({
+      next: emails => {
+        this.emailAccounts = emails;
+      },
+      error: error => {},
+    });
+
+    this.settingsService.getBucketOptions().subscribe({
+      next: storages => (this.cloudStorageList = storages),
+      error: error => {},
+    });
+  }
+
+  populateServicesSettings() {
+    const services: ServiceList[] = JSON.parse(localStorage.getItem(SERVICES));
+    from(services)
+      .pipe(
+        mergeMap(service => {
+          const serviceName = this.kebabToTitleCase(service.type);
+          return this.settingsService.getClientSettings(service.url).pipe(
+            catchError(error => {
+              this.snackBar.open(`${serviceName} ${IS_DOWN}`, CLOSE, {
+                duration: 2000,
+              });
+              return empty();
+            }),
+            map((data: ISettings) => {
+              if (data) {
+                data.serviceName = serviceName;
+                return data;
+              }
+            }),
+          );
+        }),
+        toArray(),
+      )
+      .subscribe({
+        next: (servicesSettings: ISettings[]) => {
+          for (const settings of servicesSettings) {
+            if (settings) {
+              this.clientsFormArray.push(
+                this.createClientsFormArrayItem(settings),
+              );
+            }
+          }
+        },
+        error: error => {},
+      });
+  }
+
+  createClientsFormArrayItem(settings: ISettings): FormGroup {
+    const client = this.formBuilder.group({
+      appURL: settings.appURL,
+      clientId: settings.clientId,
+      clientSecret: settings.clientSecret,
+      cloudStorageSettings: settings.cloudStorageSettings,
+      serviceName: settings.serviceName,
+      hide: true,
+    });
+
+    client.controls.appURL.disable();
+
+    return client;
   }
 
   populateForm(response) {
     this.authSettingsForm.controls.issuerUrl.setValue(response.issuerUrl);
     this.authSettingsForm.controls.communicationServerClientId.setValue(
       response.communicationServerClientId,
-    );
-  }
-
-  populateClientForm(clientResponse) {
-    this.infraSettingsForm.controls.appURL.setValue(clientResponse.appURL);
-    this.infraSettingsForm.controls.clientId.setValue(clientResponse.clientId);
-    this.infraSettingsForm.controls.clientSecret.setValue(
-      clientResponse.clientSecret,
     );
   }
 
@@ -85,16 +173,59 @@ export class SettingsComponent implements OnInit {
         this.authSettingsForm.controls.communicationServerClientId.value,
       )
       .subscribe({
-        next: response => {},
+        next: response => {
+          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+        },
+        error: error => {},
       });
+
+    if (
+      this.authSettingsForm.controls.communicationServerSystemEmailAccount.value
+    ) {
+      this.settingsService
+        .updateSystemEmailSettings(
+          this.authSettingsForm.controls.communicationServerSystemEmailAccount
+            .value,
+        )
+        .subscribe({
+          next: success => {
+            this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+          },
+          error: error => {},
+        });
+    }
   }
 
-  updateAuthClientSettings() {
-    this.settingsService.getClientUpdate(
-      localStorage.getItem(APP_URL),
-      localStorage.getItem(ISSUER_URL),
-      this.infraSettingsForm.controls.clientId.value,
-      this.infraSettingsForm.controls.clientSecret.value,
-    );
+  kebabToTitleCase(string: string) {
+    return string
+      .split('-')
+      .map(word => word.charAt(0).toUpperCase() + word.slice(1))
+      .join(' ');
+  }
+
+  updateClientSettings(client: FormGroup) {
+    const authServerURL = localStorage.getItem(ISSUER_URL);
+    const appURL = client.controls.appURL.value;
+    const clientId = client.controls.clientId.value;
+    const clientSecret = client.controls.clientSecret.value;
+    const cloudStorageSettings = client.controls.cloudStorageSettings.value;
+    this.settingsService
+      .updateClientSettings(appURL, {
+        authServerURL,
+        appURL,
+        clientId,
+        clientSecret,
+        cloudStorageSettings,
+      })
+      .subscribe({
+        next: success => {
+          this.snackBar.open(UPDATE_SUCCESSFUL, CLOSE, { duration: 2000 });
+        },
+        error: error => {
+          this.snackBar.open(ERROR_UPDATING_SERVICE_SETTINGS, CLOSE, {
+            duration: 2000,
+          });
+        },
+      });
   }
 }
